Scripts Section: Change `bat` to `cmd`
Scripts section is confusing because `bat` is now deprecated in favor of `cmd`. Changed accordingly, but also changed "The script language can be:" to "The script language macro can be:" because "cmd" is an interpreter not a scripting language. Now it's clearer the distinction between the macros names and the scripting language they are associated to.

Same changes also applied to diagram.